services/backend/requirements.txt
ansicon==1.89.0
asgiref==3.6.0
beautifulsoup4==4.11.1
blessed==1.19.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
dataclasses-json==0.5.7
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-modeltranslation==0.18.7
django-rest-multiple-models==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
facebook-sdk==3.1.0
google-auth==2.15.0
idna==3.4
inflection==0.5.1
inquirer==3.1.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-editor==1.0.4
python3-openid==3.2.0
pytz==2022.7
pyuseragents==1.0.5
readchar==4.0.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safeIO==1.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
translatepy==2.3
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
django-messages-drf
#psycopg2-binary==2.8.6
gunicorn